fix[iam]: move issuer key loading & issuer derivation to module
